move the packbit utils to core dir
